File reorganization, adding shortcut keys
