+++ A-Eye-Web-Chat-Assistant/manifest.json
@@ -49,15 +49,24 @@
   "commands": {
     "toggle-voice-control": {
       "suggested_key": {
-        "default": "Alt+Shift+1"
+        "default": "Alt+Shift+1",
+        "mac": "Alt+Shift+1"
       },
       "description": "Toggle voice browser control"
     },
     "toggle-voice-input": {
       "suggested_key": {
-        "default": "Alt+Shift+2"
+        "default": "Alt+Shift+2",
+        "mac": "Alt+Shift+2"
       },
       "description": "Toggle voice input for chat"
+    },
+    "_execute_action": {
+      "suggested_key": {
+        "default": "Alt+Shift+Q",
+        "mac": "Alt+Shift+Q"
+      },
+      "description": "Open Extension"
     }
   }
 }
