+++ A-Eye-Web-Chat-Assistant/sidepanel.js
@@ -1,7 +1,14 @@
 import { AutoProcessor, AutoTokenizer, Moondream1ForConditionalGeneration, RawImage } from './lib/transformers300.js';
+import { VoiceController } from './functions/voiceControl.js';
+import { ScreenshotController } from './functions/screenShot.js';
 
 class AIScreenReader {
     constructor() {
+        this.prompts = {
+            screenshot: 'Describe the picture in about 100 words.',
+            rollingScreenshot: 'Describe the picture in about 200 words.'
+        };
+
         this.state = {
             tokenizer: null,
             processor: null,
@@ -11,25 +18,6 @@ class AIScreenReader {
             isProcessing: false,
             currentModel: 'none',
             messages: [],
-            rollingScreenshotImages: [],
-            voice: {
-                input: {
-                    active: false,
-                    recognition: null,
-                    finalTranscript: '',
-                    silenceTimeout: null
-                },
-                control: {
-                    active: false,
-                    recognition: null
-                },
-                synthesis: {
-                    instance: window.speechSynthesis,
-                    selectedVoice: null,
-                    voices: [],
-                    isSpeaking: false
-                }
-            },
             geminiSession: null
         };
 
@@ -48,338 +36,40 @@ class AIScreenReader {
             currentModel: document.getElementById('current-model')
         };
 
-        this.voiceCommands = {
-            'take screenshot': () => this.handleScreenshot(),
-            'take rolling screenshot': () => this.handleRollingScreenshot(),
-            'analyze content': () => this.handleContentAnalysis(),
-            'analyze page': () => this.handleContentAnalysis()
-        };
+        this.voiceController = new VoiceController();
+        this.voiceController.setCallbacks({
+            appendMessage: (role, content) => this.appendMessage(role, content),
+            updateVoiceInputButtonState: (isActive) => {
+                this.elements.voiceButton.style.backgroundColor = isActive ? '#dc3545' : '#007bff';
+                this.elements.voiceButton.textContent = isActive ? 'Stop' : 'Voice';
+            },
+            handleScreenshot: () => this.handleScreenshot(),
+            handleRollingScreenshot: () => this.handleRollingScreenshot(),
+            handleContentAnalysis: () => this.handleContentAnalysis(),
+            performGoogleSearch: async (query) => {
+                try {
+                    const encodedQuery = encodeURIComponent(query);
+                    const searchUrl = `https://www.google.com/search?q=${encodedQuery}`;
+                    await chrome.tabs.create({ url: searchUrl });
+                    this.voiceController.speakText(`Searching for ${query}`);
+                    this.appendMessage('system', `Searching Google for: "${query}"`);
+                } catch (error) {
+                    console.error('Search error:', error);
+                    this.appendMessage('system', 'Failed to perform search');
+                }
+            },
+            navigateToWebsite: (website) => {
+                const url = `https://www.${website.toLowerCase()}.com`;
+                chrome.tabs.create({ url });
+                this.voiceController.speakText(`Opening ${website}`);
+            },
+            handleSendMessage: () => this.handleSendMessage()
+        });
+
+        this.screenshotController = new ScreenshotController();
 
         this.initializeAll();
         this.setupMessageListener();
-    }
-
-    initializeVoiceControl() {
-        if (!('webkitSpeechRecognition' in window)) {
-            console.error('Speech recognition not supported');
-            return;
-        }
-
-        const recognition = new webkitSpeechRecognition();
-        recognition.continuous = false;
-        recognition.interimResults = false;
-        recognition.lang = 'en-US';
-
-        recognition.onstart = () => {
-            console.log('Voice control started');
-            this.state.voice.control.active = true;
-            this.speakText('Voice control activated.');
-            this.appendMessage('system', 'Voice control activated.');
-        };
-
-        recognition.onresult = async (event) => {
-            const command = event.results[event.results.length - 1][0].transcript.toLowerCase().trim();
-            this.appendMessage('system', `Detected command: "${command}"`);
-            console.log('Voice command received:', command);
-            await this.handleVoiceCommand(command);
-
-            setTimeout(() => {
-                this.stopVoiceControl();
-            }, 800);
-        };
-
-        recognition.onerror = (event) => {
-            console.error('Voice control error:', event.error);
-            this.stopVoiceControl();
-        };
-
-        recognition.onend = () => {
-            if (this.state.voice.control.active) {
-                console.log('Voice control ended');
-                this.state.voice.control.active = false;
-            }
-        };
-
-        this.state.voice.control.recognition = recognition;
-    }
-
-    async toggleVoiceControl() {
-        if (this.state.voice.input.active) {
-            this.stopVoiceInput();
-        }
-
-        if (!this.state.voice.control.recognition) {
-            this.initializeVoiceControl();
-        }
-
-        if (this.state.voice.control.active) {
-            this.stopVoiceControl();
-        } else {
-            await this.startVoiceControl();
-        }
-    }
-
-    async startVoiceControl() {
-        try {
-            const hasPermission = await this.requestMicrophonePermission();
-            if (!hasPermission) return;
-
-            await this.state.voice.control.recognition.start();
-            this.state.voice.control.active = true;
-        } catch (error) {
-            console.error('Failed to start voice control:', error);
-            this.appendMessage('system', 'Failed to start voice control');
-            this.stopVoiceControl();
-        }
-    }
-
-    stopVoiceControl() {
-        if (!this.state.voice.control.recognition) return;
-
-        try {
-            this.state.voice.control.recognition.stop();
-            this.state.voice.control.active = false;
-        } catch (error) {
-            console.error('Error stopping voice control:', error);
-        }
-    }
-
-    async handleVoiceCommand(command) {
-        console.log('Processing command:', command);
-
-        const normalizeCommand = (cmd) => {
-            return cmd
-                .toLowerCase()
-                .replace(/analyse/g, 'analyze')
-                .replace(/screenshot/g, 'screenshot')
-                .trim();
-        };
-
-        const matchCommand = (input, target) => {
-            const normalizedInput = normalizeCommand(input);
-            return normalizedInput.includes(target) ||
-                normalizedInput.replace(/\s+/g, '') === target.replace(/\s+/g, '');
-        };
-
-        const normalizedCommand = normalizeCommand(command);
-
-        const searchMatch = normalizedCommand.match(/^search\s+(.+)$/i);
-        if (searchMatch) {
-            const searchQuery = searchMatch[1].trim();
-            await this.performGoogleSearch(searchQuery);
-            return;
-        }
-
-        const websiteMatch = normalizedCommand.match(/go to (.*?)(?:\.com|$)/i);
-        if (websiteMatch) {
-            const website = websiteMatch[1].trim();
-            await this.navigateToWebsite(website);
-            return;
-        }
-        const commandMap = {
-            'screenshot': {
-                variants: ['take screenshot', 'take a screenshot', 'capture screen'],
-                handler: async () => {
-                    this.speakText('Taking screenshot');
-                    await this.handleScreenshot();
-                }
-            },
-            'rolling': {
-                variants: ['take rolling screenshot', 'take a rolling screenshot', 'rolling screenshot'],
-                handler: async () => {
-                    this.speakText('Taking rolling screenshot');
-                    await this.handleRollingScreenshot();
-                }
-            },
-            'analyze': {
-                variants: ['analyze content', 'analyse content', 'analyze page', 'analyse page',
-                    'analyze', 'analyse', 'content analysis'],
-                handler: async () => {
-                    this.speakText('Analyzing content');
-                    await this.handleContentAnalysis();
-                }
-            }
-        };
-
-        for (const [key, { variants, handler }] of Object.entries(commandMap)) {
-            if (variants.some(variant => matchCommand(normalizedCommand, variant))) {
-                await handler();
-                return;
-            }
-        }
-
-        this.appendMessage('system', `Command not recognized: "${command}"`);
-        this.speakText("Command not recognized. Please try again.");
-    }
-
-    initializeVoiceInput() {
-        if (!('webkitSpeechRecognition' in window)) {
-            console.error('Speech recognition not supported');
-            return;
-        }
-
-        const recognition = new webkitSpeechRecognition();
-        recognition.continuous = true;
-        recognition.interimResults = true;
-        recognition.lang = 'en-US';
-
-        recognition.onstart = () => {
-            console.log('Voice input started');
-            this.state.voice.input.active = true;
-            this.updateVoiceInputButtonState(true);
-            this.elements.userInput.placeholder = 'Listening...';
-            this.state.voice.input.finalTranscript = '';
-        };
-
-        recognition.onresult = (event) => {
-            let interimTranscript = '';
-
-            if (this.state.voice.input.silenceTimeout) {
-                clearTimeout(this.state.voice.input.silenceTimeout);
-            }
-
-            for (let i = event.resultIndex; i < event.results.length; i++) {
-                const transcript = event.results[i][0].transcript;
-                if (event.results[i].isFinal) {
-                    this.state.voice.input.finalTranscript = transcript;
-                } else {
-                    interimTranscript = transcript;
-                }
-            }
-
-            this.elements.userInput.value = this.state.voice.input.finalTranscript || interimTranscript;
-
-            this.state.voice.input.silenceTimeout = setTimeout(() => {
-                if (this.state.voice.input.finalTranscript.trim()) {
-                    this.handleSendMessage();
-                    this.stopVoiceInput();
-                }
-            }, 600);
-        };
-
-        recognition.onerror = (event) => {
-            console.error('Voice input error:', event.error);
-            this.stopVoiceInput();
-        };
-
-        recognition.onend = () => {
-            console.log('Voice input ended');
-            this.stopVoiceInput();
-        };
-
-        this.state.voice.input.recognition = recognition;
-    }
-
-    async performGoogleSearch(query) {
-        try {
-            const encodedQuery = encodeURIComponent(query);
-            const searchUrl = `https://www.google.com/search?q=${encodedQuery}`;
-            await chrome.tabs.create({ url: searchUrl });
-            this.speakText(`Searching for ${query}`);
-            this.appendMessage('system', `Searching Google for: "${query}"`);
-        } catch (error) {
-            console.error('Search error:', error);
-            this.appendMessage('system', 'Failed to perform search');
-        }
-    }
-
-    async toggleVoiceInput() {
-        if (this.state.voice.control.active) {
-            this.stopVoiceControl();
-        }
-
-        if (!this.state.voice.input.recognition) {
-            this.initializeVoiceInput();
-        }
-
-        if (this.state.voice.input.active) {
-            this.stopVoiceInput();
-        } else {
-            await this.startVoiceInput();
-        }
-    }
-
-    async startVoiceInput() {
-        try {
-            const hasPermission = await this.requestMicrophonePermission();
-            if (!hasPermission) return;
-
-            await this.state.voice.input.recognition.start();
-            this.state.voice.input.active = true;
-            this.updateVoiceInputButtonState(true);
-            this.appendMessage('system', 'Voice input activated');
-            this.speakText('Voice input activated');
-        } catch (error) {
-            console.error('Failed to start voice input:', error);
-            this.stopVoiceInput();
-        }
-    }
-
-    stopVoiceInput() {
-        if (!this.state.voice.input.recognition) return;
-
-        try {
-            this.state.voice.input.recognition.stop();
-            this.state.voice.input.active = false;
-            this.updateVoiceInputButtonState(false);
-            this.elements.userInput.placeholder = 'Type your message here...';
-        } catch (error) {
-            console.error('Error stopping voice input:', error);
-        }
-    }
-
-    navigateToWebsite(website) {
-        const url = `https://www.${website.toLowerCase()}.com`;
-        chrome.tabs.create({ url });
-        this.speakText(`Opening ${website}`);
-    }
-
-    initializeSpeechSynthesis() {
-        if (!this.state.voice.synthesis.instance) return;
-
-        this.state.voice.synthesis.instance.addEventListener('voiceschanged', () => {
-            this.state.voice.synthesis.voices = this.state.voice.synthesis.instance.getVoices();
-            this.state.voice.synthesis.selectedVoice = this.state.voice.synthesis.voices.find(
-                voice => voice.lang.includes('en-US')
-            );
-        });
-    }
-
-    speakText(text) {
-        if (!this.state.voice.synthesis.instance) return;
-
-        if (this.state.voice.synthesis.isSpeaking) {
-            this.state.voice.synthesis.instance.cancel();
-        }
-
-        const utterance = new SpeechSynthesisUtterance(text);
-        utterance.voice = this.state.voice.synthesis.selectedVoice;
-        utterance.lang = 'en-US';
-        utterance.rate = 1.5;
-        utterance.pitch = 1.0;
-        utterance.volume = 1.0;
-
-        this.state.voice.synthesis.isSpeaking = true;
-        utterance.onend = () => this.state.voice.synthesis.isSpeaking = false;
-
-        this.state.voice.synthesis.instance.speak(utterance);
-    }
-
-    async requestMicrophonePermission() {
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
-            stream.getTracks().forEach(track => track.stop());
-            return true;
-        } catch (error) {
-            console.error('Microphone permission error:', error);
-            this.appendMessage('system', 'Please allow microphone access');
-            return false;
-        }
-    }
-
-    updateVoiceInputButtonState(isActive) {
-        this.elements.voiceButton.style.backgroundColor = isActive ? '#dc3545' : '#007bff';
-        this.elements.voiceButton.textContent = isActive ? 'Stop' : 'Voice';
     }
 
     setupMessageListener() {
@@ -387,8 +77,8 @@ class AIScreenReader {
             console.log('Message received in sidepanel:', request);
 
             const messageHandlers = {
-                toggleVoiceInput: () => this.toggleVoiceInput(),
-                toggleVoiceControl: () => this.toggleVoiceControl()
+                toggleVoiceInput: () => this.voiceController.toggleVoiceInput(),
+                toggleVoiceControl: () => this.voiceController.toggleVoiceControl()
             };
 
             if (messageHandlers[request.type]) {
@@ -404,9 +94,7 @@ class AIScreenReader {
             this.initializeGemini()
         ]);
         this.setupEventListeners();
-        this.initializeVoiceControl();
-        this.initializeVoiceInput();
-        this.initializeSpeechSynthesis();
+        this.voiceController.initializeAll();
     }
 
     async initializeGemini() {
@@ -424,7 +112,7 @@ class AIScreenReader {
             'screenshotButton': () => this.handleScreenshot(),
             'rollingScreenshotButton': () => this.handleRollingScreenshot(),
             'sendButton': () => this.handleSendMessage(),
-            'voiceButton': () => this.toggleVoiceInput(),
+            'voiceButton': () => this.voiceController.toggleVoiceInput(),
             'clearButton': () => this.handleClear(),
             'analyzeContentButton': () => this.handleContentAnalysis()
         };
@@ -490,7 +178,7 @@ class AIScreenReader {
 
         try {
             modelStatus.textContent = 'Model initialization in progress, please wait.';
-            this.speakText("Model initialization in progress, please wait.");
+            this.voiceController.speakText("Model initialization in progress, please wait.");
 
             const [tokenizer, processor, model] = await Promise.all([
                 AutoTokenizer.from_pretrained(modelId),
@@ -508,7 +196,7 @@ class AIScreenReader {
             Object.assign(this.state, { tokenizer, processor, model, isInitialized: true });
             this.elements.screenshotButton.disabled = false;
             modelStatus.textContent = 'Model initialization complete.';
-            this.speakText("Model initialization complete.");
+            this.voiceController.speakText("Model initialization complete.");
         } catch (error) {
             modelStatus.textContent = `Initialization failed: ${error.message}`;
             this.handleError('Error initializing model', error);
@@ -542,14 +230,14 @@ class AIScreenReader {
 
         try {
             this.elements.screenshotButton.disabled = true;
-            this.speakText("Analyzing Screenshot");
+            this.voiceController.speakText("Analyzing Screenshot");
             this.appendMessage('system', 'Analyzing Screenshot.');
             this.updateModel('moondream');
 
-            const screenshot = await chrome.tabs.captureVisibleTab();
+            const screenshot = await this.screenshotController.captureVisibleTab();
             this.showPreview('image', screenshot);
 
-            await this.handleImageAnalysis(screenshot);
+            await this.handleImageAnalysis(screenshot, this.prompts.screenshot);
         } catch (error) {
             this.handleError('Error taking screenshot', error);
         } finally {
@@ -562,24 +250,17 @@ class AIScreenReader {
 
         try {
             this.elements.rollingScreenshotButton.disabled = true;
-            this.state.rollingScreenshotImages = [];
-            this.speakText("Analyzing scrolling screenshot.");
-            this.appendMessage('system', 'Analyzing scrolling screenshot.');
+            this.voiceController.speakText("Analyzing rolling screenshot.");
+            this.appendMessage('system', 'Analyzing rolling screenshot.');
             this.updateModel('moondream');
 
             const [tab] = await chrome.tabs.query({ active: true, currentWindow: true });
             if (!tab) throw new Error('No active tab found');
 
-            const pageInfo = await this.getPageInfo();
-            if (!this.validatePageInfo(pageInfo)) {
-                throw new Error('Invalid page dimensions');
-            }
-
-            await this.captureScreenshots(tab, pageInfo);
-            const mergedImage = await this.mergeScreenshots(this.state.rollingScreenshotImages);
+            const mergedImage = await this.screenshotController.handleRollingScreenshot(tab);
             this.showPreview('image', mergedImage);
 
-            await this.handleImageAnalysis(mergedImage);
+            await this.handleImageAnalysis(mergedImage, this.prompts.rollingScreenshot);
         } catch (error) {
             this.handleError('Rolling screenshot failed', error);
         } finally {
@@ -587,87 +268,7 @@ class AIScreenReader {
         }
     }
 
-    validatePageInfo(pageInfo) {
-        return pageInfo && pageInfo.scrollHeight && pageInfo.clientHeight;
-    }
-
-    async getPageInfo() {
-        return new Promise((resolve, reject) => {
-            chrome.runtime.sendMessage({
-                type: 'startRollingScreenshot'
-            }, response => {
-                if (!response) {
-                    reject(new Error('Failed to get page dimensions'));
-                    return;
-                }
-                resolve(response);
-            });
-        });
-    }
-
-    async captureScreenshots(tab, pageInfo) {
-        const { scrollHeight, clientHeight } = pageInfo;
-        let currentScroll = 0;
-
-        await this.executeScroll(tab, 0);
-
-        while (currentScroll < scrollHeight) {
-            try {
-                const screenshot = await chrome.tabs.captureVisibleTab(null, { format: 'png' });
-                this.state.rollingScreenshotImages.push(screenshot);
-            } catch (error) {
-                if (error.message.includes('MAX_CAPTURE_VISIBLE_TAB_CALLS_PER_SECOND')) {
-                    console.warn('Rate limit exceeded. Waiting before retrying...');
-                    await new Promise(resolve => setTimeout(resolve, 1000));
-                    continue;
-                } else {
-                    throw error;
-                }
-            }
-
-            currentScroll += clientHeight;
-            await this.executeScroll(tab, currentScroll);
-
-            await new Promise(resolve => setTimeout(resolve, 600));
-        }
-
-        await this.executeScroll(tab, 0);
-    }
-
-    async executeScroll(tab, scrollPosition) {
-        await chrome.scripting.executeScript({
-            target: { tabId: tab.id },
-            function: (scroll) => window.scrollTo({ top: scroll, behavior: 'instant' }),
-            args: [scrollPosition]
-        });
-    }
-
-    async mergeScreenshots(screenshots) {
-        const canvas = document.createElement('canvas');
-        const ctx = canvas.getContext('2d');
-
-        const firstImage = await this.loadImage(screenshots[0]);
-        canvas.width = firstImage.width;
-        canvas.height = firstImage.height * screenshots.length;
-
-        for (let i = 0; i < screenshots.length; i++) {
-            const image = await this.loadImage(screenshots[i]);
-            ctx.drawImage(image, 0, i * firstImage.height);
-        }
-
-        return canvas.toDataURL('image/png');
-    }
-
-    loadImage(src) {
-        return new Promise((resolve, reject) => {
-            const img = new Image();
-            img.onload = () => resolve(img);
-            img.onerror = reject;
-            img.src = src;
-        });
-    }
-
-    async handleImageAnalysis(imageUrl) {
+    async handleImageAnalysis(imageUrl, prompt = this.prompts.screenshot) {
         if (!this.state.isInitialized || this.state.isProcessing) {
             this.handleError('Model not ready', new Error('Please wait for model initialization'));
             return;
@@ -677,9 +278,7 @@ class AIScreenReader {
 
         try {
             this.state.messages = [];
-            const defaultPrompt = 'Describe the picture in about 100 words';
-
-            const response = await this.generateImageResponse(imageUrl, defaultPrompt);
+            const response = await this.generateImageResponse(imageUrl, prompt);
             this.handleResponse(response);
         } catch (error) {
             this.handleError('Error analyzing screenshot', error);
@@ -743,7 +342,7 @@ class AIScreenReader {
 
             this.elements.userInput.value = '';
         } catch (error) {
-            this.handleError('Error sending message', error);
+            this.handleError('Error', error);
         } finally {
             this.state.isProcessing = false;
             this.enableInterface();
@@ -762,7 +361,7 @@ class AIScreenReader {
             this.updateModel('gemini');
             this.state.isProcessing = true;
             this.elements.analyzeContentButton.disabled = true;
-            this.speakText("Analyzing webpage content.");
+            this.voiceController.speakText("Analyzing webpage content.");
             this.appendMessage('assistant', 'Analyzing webpage content.');
 
             const [tab] = await chrome.tabs.query({ active: true, currentWindow: true });
@@ -865,7 +464,7 @@ class AIScreenReader {
     handleResponse(response) {
         this.appendMessage('assistant', response);
         this.state.messages.push({ role: 'assistant', content: response });
-        this.speakText(response);
+        this.voiceController.speakText(response);
         this.enableInterface();
     }
 
@@ -877,8 +476,6 @@ class AIScreenReader {
 
     resetState() {
         this.state.isProcessing = false;
-        this.state.voice.input.active = false;
-        this.state.rollingScreenshotImages = [];
         this.enableInterface();
     }
 
@@ -889,17 +486,13 @@ class AIScreenReader {
 
         Object.assign(this.state, {
             messages: [],
-            rollingScreenshotImages: [],
             pastKeyValues: null,
             currentModel: 'none'
         });
 
         this.elements.currentModel.textContent = 'None';
-
-        if (this.state.voice.synthesis.instance && this.state.voice.synthesis.isSpeaking) {
-            this.state.voice.synthesis.instance.cancel();
-            this.state.voice.synthesis.isSpeaking = false;
-        }
+        this.voiceController.cleanup();
+        this.screenshotController.cleanup();
     }
 
     enableInterface() {
@@ -916,3 +509,7 @@ class AIScreenReader {
 }
 
 const aiScreenReader = new AIScreenReader();
+
+document.addEventListener('DOMContentLoaded', function () {
+    console.log('Extension loaded!');
+});
